1) Moved all the MasterDataEditorGMF actions in "File -> MasterDataEditorGMF diagram".
2) The default location for all the MasterDataEditorGMF files is now
"C:\<Documents and Settings path>\AspireRFID\IDE\BPWME".

3) For "New MasterDataEditorGMF Diagram" and "Open MasterDataEditorGMF
Diagram From Epcis" actions, the user is prompted to select a CLCBProc (of
open BPWME diagrams) in order to correlate it with the MasterData. The
name of the MasterDataEditor file is set to "<CLCBProc
name>.masterdataeditorgmf_diagram".
If no selection is made the file takes its default name.

4) For "Open MasterDataEditorGMF Diagram From File" and "Open
MasterDataEditorGMF Diagram From Apdl" actions, if the file to be opened
does not have the extension "masterdataeditorgmf_diagram" and "xml"
respectively an error is displayed.

5) For "Open MasterDataEditorGMF Diagram From Apdl" action, the number of
MasterDataEditorGMF diagrams that are created is equal to the number of
MasterData that are included in each CLCBProc inside the Apdl file. Each
MasterDataEditorGMF filename is set to "<parent CLCBProc
name>.masterdataeditorgmf_diagram".

6) In the Configurator Preferences, the user can choose what the size of
the elements will be, when the "MasterDataEditorGMF Square Layout" action
is pressed.

7) In the Configurator Preferences, the user is prompted to restart the
application in order to set the new custom properties.